feat(wallet): add spending password ui
