+++ apps/extension/package.json
@@ -31,7 +31,7 @@
     "zod": "^3.21.4"
   },
   "devDependencies": {
-    "@crxjs/vite-plugin": "^2.0.0-beta.18",
+    "@crxjs/vite-plugin": "^2.0.0-beta.21",
     "@esbuild-plugins/node-globals-polyfill": "^0.2.3",
     "@esbuild-plugins/node-modules-polyfill": "^0.2.2",
     "@originjs/vite-plugin-commonjs": "^1.0.3",
@@ -45,17 +45,17 @@
     "@types/webextension-polyfill": "^0.10.0",
     "@typescript-eslint/eslint-plugin": "^5.57.1",
     "@typescript-eslint/parser": "^5.57.1",
-    "@vitejs/plugin-react-swc": "^3.3.2",
+    "@vitejs/plugin-react-swc": "^3.5.0",
     "path": "^0.12.7",
     "rollup-plugin-node-polyfills": "^0.2.1",
     "rollup-plugin-polyfill-node": "^0.12.0",
-    "vite": "^4.3.9",
-    "vite-plugin-commonjs": "^0.7.1",
-    "vite-plugin-node-polyfills": "^0.9.0",
-    "vite-plugin-require-transform": "^1.0.20",
-    "vite-plugin-svgr": "^3.2.0",
-    "vite-plugin-top-level-await": "^1.3.1",
-    "vite-plugin-wasm": "^3.2.2",
-    "web-ext": "^7.6.2"
+    "vite": "^4.5.1",
+    "vite-plugin-commonjs": "^0.10.1",
+    "vite-plugin-node-polyfills": "^0.17.0",
+    "vite-plugin-require-transform": "^1.0.21",
+    "vite-plugin-svgr": "^4.2.0",
+    "vite-plugin-top-level-await": "^1.4.1",
+    "vite-plugin-wasm": "^3.3.0",
+    "web-ext": "^7.9.0"
   }
 }
