+++ packages/features/package.json
@@ -112,7 +112,7 @@
     "graphql-request": "^6.1.0",
     "mina-signer": "^2.1.1",
     "swr": "^2.1.5",
-    "vite": "^4.3.9",
+    "vite": "^4.5.1",
     "vite-plugin-svgr": "^3.2.0",
     "vite-plugin-top-level-await": "^1.3.1"
   },
