+++ packages/features/src/send/components/ConfirmTransactionForm.schema.ts
@@ -1,0 +1,5 @@
+import { z } from 'zod'
+
+export const ConfirmTransactionSchema = z.object({
+  spendingPassword: z.string().min(1)
+})
