+++ packages/persistence/package.json
@@ -24,7 +24,7 @@
   "devDependencies": {
     "@capacitor/core": "^5.2.0",
     "@palladxyz/common": "*",
-    "vite": "^4.3.9",
+    "vite": "^4.5.1",
     "zustand": "^4.4.7"
   }
 }
